requirements.txt
Django==5.0.1
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0
supabase==2.3.4
speechrecognition==3.10.1
pyaudio==0.2.11
pillow==10.2.0
django-cors-headers==4.3.1
whitenoise==6.6.0